Fixed mouse click reporting (pass it as h instead of ch)

+++ src/termbox.h
@@ -132,10 +132,10 @@ struct tb_event {
 	uint16_t key; /* one of the TB_KEY_* constants */
 	uint32_t ch; /* unicode character */
 	uint8_t meta;
-	int32_t w;
-	int32_t h;
-	int32_t x;
-	int32_t y;
+	int16_t w;
+	int16_t h;
+	int16_t x;
+	int16_t y;
 };
 
 /* Error codes returned by tb_init(). All of them are self-explanatory, except
